FlaskProject/requirements.txt
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
computer-vision==0.0.6
cycler==0.10.0
decorator==4.4.2
Flask==1.1.2
idna==2.10
imageio==2.9.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
networkx==2.5
numpy==1.19.4
opencv-python==4.4.0.46
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
requests==2.25.0
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
tifffile==2020.11.26
urllib3==1.26.2
validators==0.18.1
Werkzeug==1.0.1
